+++ common/src/main/java/me/brynview/navidrohim/jmws/Constants.java
@@ -16,9 +16,12 @@ public class Constants {
     }
 
     public static final String MODID = "jmws";
-    public static final String VERSION = "1.1.6-1.21.8-beta.rc2";
+    public static final String VERSION = "1.1.6-1.21.8-beta.rc3"; // This is purely for display and is not needed
 
-    public static final List<String> forgeModLoaders = List.of("Forge", "NeoForge");
-    public static final List<String> forbiddenGroups = List.of("journeymap_death", "journeymap_all", "journeymap_temp", "journeymap_default");
+    public static final List<String> forgeModLoaders = List.of("Forge", "NeoForge"); // Do not edit unless there is another fork of Forge (would not be surprised)
+    public static final List<String> forbiddenGroups = List.of("journeymap_death", "journeymap_all", "journeymap_temp", "journeymap_default"); // (DO NOT EDIT, game will bug out)
+
+    // For anyone forking, this MUST be updated if there is a change that is in a future version of JM that you use.
     public static final String JourneyMapVersionString = "1.21.8-6.0.0-beta.52";
+    // beta 52 fixed the waypoint-drag-drop event, so this is the only version compatible (and any newer)
 }